+++ docs/source/changelog.rst
@@ -3,8 +3,12 @@
 Paver Changelog
 ===============
 
+1.x.y (TBA)
+---------------------
+* new :func:`tasks.consume_nargs` decorator, similar to :func:`tasks.consume_args` but accepting an argument: the number of arguments that the decorated function will consume. If no argument is passed to ``consume_nargs`` decorator, all comand-line arguments will be consumed.
+
 1.1.0 (July 30, 2012)
-------------------------
+---------------------
 * Minilib is now imported only if full paver is not available (github `issue #13 <https://github.com/paver/paver/issues/13>`_)
 * Option instances may now be passed to @cmdopts (github issues `#41 <https://github.com/paver/paver/issues/41>`_ and `#42 <https://github.com/paver/paver/issues/42>`_, thanks to David Cramer)
 * ``--propagate-traceback`` option for debugging ``BuildFailure``s (github issue `#43 <https://github.com/paver/paver/issues/43>`_)